deactivateTabs
