+++ extension/background.js
@@ -52,143 +52,12 @@
 //       `,
 // }
 
-const pathToWhere = (start, ...path) => {
-  const pckg = typeof (start) === 'string' ? { type_id: { _eq: 2 }, string: { value: { _eq: start } } } : { id: { _eq: start } };
-  let where = pckg;
-  for (let p = 0; p < path.length; p++) {
-    const item = path[p];
-    if (typeof (item) !== 'boolean') {
-      const nextWhere = { in: { type_id: { _eq: 3 }, string: { value: { _eq: item } }, from: where } };
-      where = nextWhere;
-    }
-  }
-  return where;
-}
+import { executeInsertTabs } from "./insert-tabs";
+import { executeDeactivateTabs } from "./deactivate-tabs";
 
-const getDeepId = async (start, ...path) => {
-  const containWhere = pathToWhere(start, ...path);
-
-  const deepquery2 = {
-    "query": `query deepid($where:links_bool_exp!){links(where:$where){id}}`,
-    variables: { where: containWhere }
-  }
-
-  const response = await fetch("http://127.0.0.1:3006/gql", {
-    method: "POST", // *GET, POST, PUT, DELETE, etc.
-    // mode: "cors", // no-cors, *cors, same-origin
-    cache: "no-cache", // *default, no-cache, reload, force-cache, only-if-cached
-    // credentials: "same-origin", // include, *same-origin, omit
-    headers: {
-      "Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJodHRwczovL2hhc3VyYS5pby9qd3QvY2xhaW1zIjp7IngtaGFzdXJhLWFsbG93ZWQtcm9sZXMiOlsiYWRtaW4iXSwieC1oYXN1cmEtZGVmYXVsdC1yb2xlIjoiYWRtaW4iLCJ4LWhhc3VyYS11c2VyLWlkIjoiMzc2In0sImlhdCI6MTY3OTkzNTMyOH0.LmDNulSTSSrm7gKno3E1sBLAhz5TKi-SFBl9oFNfs-k"
-      // 'Content-Type': 'application/x-www-form-urlencoded',
-    },
-    redirect: "follow", // manual, *follow, error
-    referrerPolicy: "no-referrer", // no-referrer, *no-referrer-when-downgrade, origin, origin-when-cross-origin, same-origin, strict-origin, strict-origin-when-cross-origin, unsafe-url
-    body: JSON.stringify(deepquery2), // body data type must match "Content-Type" header
-  });
-  const result = await response.json()
-
-  console.log(start, ...path)
-  console.log(deepquery2.query);
-  console.log(result);
-  return result.data.links[0].id
-}
-
-// Define the InsertTabs mutation in GraphQL syntax
-const insertTabsMutation = `
-mutation InsertLinks($links: [links_insert_input!]!) {
-  insert_links(objects: $links) {
-    returning {
-      id
-    }
-  }
-}
-`;
-
-// Prepare the variables object using the tabs array
-const prepareInsertTabsVariables = async (tabs) => {
-  const containTypeLinkId = await getDeepId("@deep-foundation/core", "Contain");
-  const browserExtensionLinkId = await getDeepId(2457, "BrowserExtension");
-  const tabTypeLinkId = await getDeepId("@deep-foundation/browser-extension", "Tab");
-  const urlTypeLinkId = await getDeepId("@deep-foundation/browser-extension", "TabUrl");
-  const titleTypeLinkId = await getDeepId("@deep-foundation/browser-extension", "TabTitle");
-  const activeTypeLinkId = await getDeepId("@deep-foundation/browser-extension", "Active");
-  const links = tabs.map((tab) => ({
-    type_id: tabTypeLinkId,
-    number: { data: { value: tab.id } },
-    in: {
-      data: [
-        {
-          type_id: containTypeLinkId,
-          from_id: browserExtensionLinkId,
-        },
-      ],
-    },
-    out: {
-      data: [
-        {
-          type_id: containTypeLinkId,
-          to: {
-            data: {
-              type_id: urlTypeLinkId,
-              string: { data: { value: tab.url } },
-            },
-          },
-        },
-        {
-          type_id: containTypeLinkId,
-          to: {
-            data: {
-              type_id: titleTypeLinkId,
-              string: { data: { value: tab.title } },
-            },
-          },
-        },
-        {
-          type_id: containTypeLinkId,
-          to: {
-            data: {
-              type_id: activeTypeLinkId,
-              string: { data: { value: tab.active ? "true" : "false" } },
-            },
-          },
-        },
-      ],
-    },
-
-  }));
-
-  return { links };
-};
-
-// Function to execute the InsertTabs mutation
-const executeInsertTabs = async (tabs) => {
-  const variables = await prepareInsertTabsVariables(tabs);
-
-  const requestPayload = {
-    query: insertTabsMutation,
-    variables,
-  };
-
-  const response = await fetch("http://127.0.0.1:3006/gql", {
-    method: "POST",
-    cache: "no-cache",
-    headers: {
-      "Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJodHRwczovL2hhc3VyYS5pby9qd3QvY2xhaW1zIjp7IngtaGFzdXJhLWFsbG93ZWQtcm9sZXMiOlsiYWRtaW4iXSwieC1oYXN1cmEtZGVmYXVsdC1yb2xlIjoiYWRtaW4iLCJ4LWhhc3VyYS11c2VyLWlkIjoiMzc2In0sImlhdCI6MTY3OTkzNTMyOH0.LmDNulSTSSrm7gKno3E1sBLAhz5TKi-SFBl9oFNfs-k",
-      "Content-Type": "application/json",
-    },
-    redirect: "follow",
-    referrerPolicy: "no-referrer",
-    body: JSON.stringify(requestPayload),
-  });
-
-  const responseData = await response.json();
-  console.log(requestPayload.query);
-  console.log(requestPayload.variables);
-  console.log(responseData);
-
-  return responseData.data;
-};
+// Usage example:
+const tabLinkId = "your-tab-link-id";
+executeDeleteLinks(tabLinkId);
 
 chrome.runtime.onInstalled.addListener((reason) => {
   console.log("Extension installed or updated:", reason);
@@ -200,6 +69,7 @@ chrome.runtime.onInstalled.addListener((reason) => {
   chrome.tabs.onActivated.addListener((activeInfo) => {
     // Add code to execute when a tab is activated
     console.log("Tab activated:", activeInfo.tabId);
+    executeDeactivateTabs();
   });
 
   chrome.tabs.onCreated.addListener((tab) => {
